requirements.txt
gradio
Pillow
numpy
websocket-client
fastapi 
uvicorn
python-dotenv
psycopg2